moving user list table for managers
