+++ src/main/resources/i18n/messages.properties
@@ -2468,8 +2468,7 @@ server.AdminUsersService.enabled={0} has been enabled
 server.AdminUsersService.disabled={0} has been disabled
 server.AdminUsersService.error=Error updating {0}
 
-UsersPage.title=Users
-UsersPage.add=Add User
+UserListPage.title=Users
 
 AdminUsersTable.lastLogin=Last Login
 AdminUsersTable.username=Username
