+++ src/main/webapp/entries.js
@@ -30,7 +30,7 @@ module.exports = {
   "sample-files": "./resources/js/pages/samples/sample-files.js",
   "project-analyses": "./resources/js/pages/projects/project-analyses/",
   analyses: "./resources/js/pages/analyses/AnalysesPage.jsx",
-  "users-list": "./resources/js/pages/UsersPage.jsx",
+  "users-list": "./resources/js/pages/user/components/UserListPage.jsx",
   "users-password": "./resources/js/pages/users/users-password.js",
   "sample-edit": "./resources/js/pages/samples/sample-edit.js",
   "sample-files-concatenate":
