+++ src/test/java/ca/corefacility/bioinformatics/irida/ria/integration/pages/user/UserListPage.java
@@ -1,4 +1,4 @@
-package ca.corefacility.bioinformatics.irida.ria.integration.pages;
+package ca.corefacility.bioinformatics.irida.ria.integration.pages.user;
 
 import java.util.List;
 
@@ -8,35 +8,35 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 
 import ca.corefacility.bioinformatics.irida.ria.integration.components.AntTable;
+import ca.corefacility.bioinformatics.irida.ria.integration.pages.AbstractPage;
 
 /**
- * UI Testing for the UsersPage react component.
+ * UI Testing for the UserListPage react component.
  */
-public class UsersPage extends AbstractPage {
+public class UserListPage extends AbstractPage {
 	private static AntTable table;
 
 	@FindBy(css = ".t-cb-enable input[type='checkbox']")
 	private List<WebElement> enableCheckboxes;
 
-	public UsersPage(WebDriver driver) {
+	public UserListPage(WebDriver driver) {
 		super(driver);
 	}
 
-	public static UsersPage goTo(WebDriver driver) {
+	public static UserListPage goTo(WebDriver driver) {
 		get(driver, "users");
 		table = AntTable.getTable(driver);
-		return PageFactory.initElements(driver, UsersPage.class);
+		return PageFactory.initElements(driver, UserListPage.class);
 	}
 
-	public static UsersPage goToAdminPanel(WebDriver driver) {
+	public static UserListPage goToAdminPanel(WebDriver driver) {
 		get(driver, "admin/users");
 		table = AntTable.getTable(driver);
-		return PageFactory.initElements(driver, UsersPage.class);
+		return PageFactory.initElements(driver, UserListPage.class);
 	}
 
 	public int usersTableSize() {
-		return table.getRows()
-				.size();
+		return table.getRows().size();
 	}
 
 	public boolean isTableSortedByUsername() {
@@ -56,7 +56,6 @@ public class UsersPage extends AbstractPage {
 	}
 
 	public boolean canUserModifyUserState() {
-		return enableCheckboxes.get(0)
-				.isEnabled();
+		return enableCheckboxes.get(0).isEnabled();
 	}
 }
